move APIFnCommit to action_commit.go
